% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goose_survey_tools.R
\name{abbreviate_survey_question}
\alias{abbreviate_survey_question}
\title{Intelligent Abbreviation of Survey Questions}
\usage{
abbreviate_survey_question(question, max_length = 20, custom_dict = NULL)
}
\arguments{
\item{question}{The survey question text}

\item{max_length}{Maximum length for the abbreviated name}

\item{custom_dict}{Custom abbreviation dictionary}
}
\value{
Abbreviated variable name
}
\description{
Internal function that intelligently abbreviates survey questions
into short, meaningful variable names.
}
\keyword{internal}
