% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{brand_css_inline}
\alias{brand_css_inline}
\title{Get inline CSS for RMarkdown}
\usage{
brand_css_inline(brand = "block")
}
\arguments{
\item{brand}{Brand name}
}
\value{
HTML style tag with CSS
}
\description{
Helper function to embed CSS directly in RMarkdown
}
\keyword{internal}
