% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{brand_palette}
\alias{brand_palette}
\title{Get brand color palette}
\usage{
brand_palette(brand = "block", palette = "categorical", n = NULL)
}
\arguments{
\item{brand}{Name of the brand}

\item{palette}{Type of palette ("categorical", "sequential", "diverging")}

\item{n}{Number of colors to return (NULL for all)}
}
\value{
Character vector of hex colors
}
\description{
Get brand color palette
}
\examples{
\dontrun{
colors <- brand_palette("block", "categorical")
}
}
