% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_integration.R
\name{extract_session_id}
\alias{extract_session_id}
\title{Extract session ID from Goose output}
\usage{
extract_session_id(output)
}
\description{
Extract session ID from Goose output
}
\keyword{internal}
