% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_conversation}
\alias{format_conversation}
\title{Format and display a gooseR conversation}
\usage{
format_conversation(messages, show_dividers = TRUE)
}
\arguments{
\item{messages}{List of messages (question/response pairs)}

\item{show_dividers}{Logical, show dividers between messages}
}
\description{
Format and display a gooseR conversation
}
