% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{generate_diverging_palette}
\alias{generate_diverging_palette}
\title{Generate diverging color palette}
\usage{
generate_diverging_palette(color1, color2, n = 5)
}
\description{
Generate diverging color palette
}
\keyword{internal}
