% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{generate_sequential_palette}
\alias{generate_sequential_palette}
\title{Generate sequential color palette}
\usage{
generate_sequential_palette(base_color, n = 5)
}
\description{
Generate sequential color palette
}
\keyword{internal}
