% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{goose_batch}
\alias{goose_batch}
\title{Execute Multiple Queries in Parallel}
\usage{
goose_batch(queries, max_workers = 4, progress = TRUE, cache = TRUE)
}
\arguments{
\item{queries}{Vector or list of queries}

\item{max_workers}{Maximum number of parallel workers}

\item{progress}{Show progress bar}

\item{cache}{Use caching for responses}
}
\value{
List of responses
}
\description{
Run multiple queries simultaneously
}
\examples{
\dontrun{
queries <- c(
  "Explain R functions",
  "Write a data analysis script",
  "Create a visualization"
)

results <- goose_batch(queries, max_workers = 3)
}
}
