% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{goose_batch_file}
\alias{goose_batch_file}
\title{Batch Process File}
\usage{
goose_batch_file(file, output_file = NULL, max_workers = 4)
}
\arguments{
\item{file}{Path to file with queries (one per line)}

\item{output_file}{Optional output file for results}

\item{max_workers}{Maximum parallel workers}
}
\value{
List of results
}
\description{
Process a file of queries in parallel
}
