% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{goose_cache_clear}
\alias{goose_cache_clear}
\title{Clear Cache}
\usage{
goose_cache_clear(older_than = NULL, pattern = NULL, all = FALSE, conn = NULL)
}
\arguments{
\item{older_than}{Clear entries older than this (seconds)}

\item{pattern}{Clear entries matching this pattern}

\item{all}{Clear all entries}

\item{conn}{Database connection}
}
\value{
Number of entries cleared
}
\description{
Clear cache entries based on criteria
}
