% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{goose_cache_export}
\alias{goose_cache_export}
\title{Export Cache}
\usage{
goose_cache_export(file, format = "json", conn = NULL)
}
\arguments{
\item{file}{Path to export file}

\item{format}{Export format (json, csv, rds)}

\item{conn}{Database connection}
}
\value{
Number of entries exported
}
\description{
Export cache to file for backup or sharing
}
