% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{goose_cache_import}
\alias{goose_cache_import}
\title{Import Cache}
\usage{
goose_cache_import(file, format = "json", conn = NULL)
}
\arguments{
\item{file}{Path to import file}

\item{format}{Import format (json, csv, rds)}

\item{conn}{Database connection}
}
\value{
Number of entries imported
}
\description{
Import cache from file
}
