% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{goose_cache_stats}
\alias{goose_cache_stats}
\title{Get Cache Statistics}
\usage{
goose_cache_stats(conn = NULL)
}
\arguments{
\item{conn}{Database connection}
}
\value{
List of cache statistics
}
\description{
Get statistics about cache usage
}
