% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{goose_cache_warmup}
\alias{goose_cache_warmup}
\title{Cache Warmup}
\usage{
goose_cache_warmup(queries, parallel = FALSE)
}
\arguments{
\item{queries}{Vector of queries to cache}

\item{parallel}{Execute in parallel}
}
\value{
Number of queries cached
}
\description{
Pre-populate cache with common queries
}
