% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{goose_cached}
\alias{goose_cached}
\title{Cached Query Execution}
\usage{
goose_cached(query, use_cache = TRUE, max_age = 86400, force_refresh = FALSE)
}
\arguments{
\item{query}{The query to execute}

\item{use_cache}{Whether to use cache}

\item{max_age}{Maximum cache age in seconds}

\item{force_refresh}{Force new execution}
}
\value{
Query response
}
\description{
Execute query with automatic caching
}
