% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_integration.R
\name{goose_check_installation}
\alias{goose_check_installation}
\title{Check Goose CLI Installation}
\usage{
goose_check_installation()
}
\value{
Logical, TRUE if Goose CLI is installed
}
\description{
Check Goose CLI Installation
}
