% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{goose_create_brand}
\alias{goose_create_brand}
\title{Create a new brand configuration interactively}
\usage{
goose_create_brand(brand_name = NULL, template = "default", interactive = TRUE)
}
\arguments{
\item{brand_name}{Name for the new brand}

\item{template}{Template to use (default: "default")}

\item{interactive}{Use interactive prompts (default: TRUE)}
}
\value{
Path to created brand configuration
}
\description{
Create a new brand configuration interactively
}
\examples{
\dontrun{
goose_create_brand("my_company")
}
}
