% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_integration.R
\name{goose_create_template}
\alias{goose_create_template}
\title{Create Parameterized Report Template}
\usage{
goose_create_template(
  type = "analysis",
  parameters = list(),
  output_file = NULL
)
}
\arguments{
\item{type}{Report type (analysis, dashboard, presentation)}

\item{parameters}{List of parameters}

\item{output_file}{Output file path}
}
\value{
Path to template file
}
\description{
Generate a parameterized Quarto/RMarkdown template
}
