% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{goose_divider}
\alias{goose_divider}
\title{Create a markdown-style divider}
\usage{
goose_divider(char = "─", width = 60)
}
\arguments{
\item{char}{Character to use for divider}

\item{width}{Width of divider}
}
\description{
Create a markdown-style divider
}
