% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_assistant.R
\name{goose_generate_tests}
\alias{goose_generate_tests}
\title{Generate Unit Tests with AI}
\usage{
goose_generate_tests(func, test_cases = 5, edge_cases = TRUE)
}
\arguments{
\item{func}{Function object to test}

\item{test_cases}{Number of test cases to generate}

\item{edge_cases}{Logical, include edge case tests}
}
\value{
Character string with testthat code
}
\description{
Generate testthat unit tests for R functions.
}
