% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_bulk.R
\name{goose_rename}
\alias{goose_rename}
\title{Rename an item in gooseR memory}
\usage{
goose_rename(old_name, new_name, category = "general")
}
\arguments{
\item{old_name}{Character string, current name}

\item{new_name}{Character string, new name}

\item{category}{Character string, category of the item}
}
\value{
Logical, TRUE if successful
}
\description{
Rename an item in gooseR memory
}
\examples{
\dontrun{
goose_rename("old_analysis", "final_analysis", category = "results")
}
}
