% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_bulk.R
\name{goose_session_clear}
\alias{goose_session_clear}
\title{Clear all items from current session}
\usage{
goose_session_clear(confirm = TRUE)
}
\arguments{
\item{confirm}{Logical, whether to confirm deletion}
}
\value{
Invisible integer count of deleted items
}
\description{
Clear all items from current session
}
