% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_bulk.R
\name{goose_session_end}
\alias{goose_session_end}
\title{End the current gooseR session}
\usage{
goose_session_end(cleanup = FALSE)
}
\arguments{
\item{cleanup}{Logical, whether to delete session items}
}
\value{
Invisible NULL
}
\description{
End the current gooseR session
}
