% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_bulk.R
\name{goose_session_list}
\alias{goose_session_list}
\title{List items saved in current session}
\usage{
goose_session_list()
}
\value{
Data frame of session items
}
\description{
List items saved in current session
}
