% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_bulk.R
\name{goose_session_save}
\alias{goose_session_save}
\title{Save an object with session tracking}
\usage{
goose_session_save(..., tags = NULL)
}
\arguments{
\item{...}{Arguments passed to goose_save}

\item{tags}{Additional tags (session tag will be added automatically)}
}
\value{
Same as goose_save
}
\description{
A wrapper around goose_save that automatically adds the current session tag
}
