% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{goose_stream_multi}
\alias{goose_stream_multi}
\title{Stream Multiple Queries}
\usage{
goose_stream_multi(queries, callback = NULL)
}
\arguments{
\item{queries}{Vector of queries to execute}

\item{callback}{Optional callback for each response}
}
\value{
List of responses
}
\description{
Stream multiple queries sequentially with progress
}
