% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{goose_template_apply}
\alias{goose_template_apply}
\title{Apply Template with Variables}
\usage{
goose_template_apply(template, ..., execute = TRUE)
}
\arguments{
\item{template}{Template object or name}

\item{...}{Variable values}

\item{execute}{Whether to execute the query}
}
\value{
Filled template or query response
}
\description{
Fill in template variables and execute query
}
\examples{
\dontrun{
# Use the template
result <- goose_template_apply(
  review_template,
  language = "R",
  code = "function(x) x^2",
  focus = "efficiency and style"
)
}
}
