% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{goose_template_builtin}
\alias{goose_template_builtin}
\title{Get Built-in Template}
\usage{
goose_template_builtin(name = NULL)
}
\arguments{
\item{name}{Optional template name}
}
\value{
Template or list of templates
}
\description{
Access built-in template library
}
