% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{goose_template_from_query}
\alias{goose_template_from_query}
\title{Create Template from History}
\usage{
goose_template_from_query(query, name, variables = NULL)
}
\arguments{
\item{query}{Previous query text}

\item{name}{Template name}

\item{variables}{Variables to extract}
}
\value{
Template object
}
\description{
Create a template from a previous query
}
