% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{goose_template_list}
\alias{goose_template_list}
\title{List Available Templates}
\usage{
goose_template_list(include_builtin = TRUE)
}
\arguments{
\item{include_builtin}{Include built-in templates}
}
\value{
Data frame of templates
}
\description{
Show all available templates
}
