% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{goose_template_load}
\alias{goose_template_load}
\title{Load Template from Library}
\usage{
goose_template_load(name)
}
\arguments{
\item{name}{Template name}
}
\value{
Template object
}
\description{
Load a saved template
}
