% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{goose_template_save}
\alias{goose_template_save}
\title{Save Template to Library}
\usage{
goose_template_save(template, overwrite = FALSE)
}
\arguments{
\item{template}{Template object}

\item{overwrite}{Overwrite if exists}
}
\value{
Logical indicating success
}
\description{
Save a template for future use
}
