% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{goose_template_validate}
\alias{goose_template_validate}
\title{Validate Template}
\usage{
goose_template_validate(template)
}
\arguments{
\item{template}{Template object or text}
}
\value{
List with validation results
}
\description{
Check if template is valid
}
