% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{goose_worker_pool}
\alias{goose_worker_pool}
\title{Create Async Worker Pool}
\usage{
goose_worker_pool(n_workers = 4)
}
\arguments{
\item{n_workers}{Number of workers}
}
\value{
WorkerPool object
}
\description{
Create a pool of workers for processing queries
}
