% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{load_brand}
\alias{load_brand}
\title{Load brand configuration}
\usage{
load_brand(brand = "block")
}
\arguments{
\item{brand}{Name of the brand to load}
}
\value{
List containing brand configuration
}
\description{
Load brand configuration
}
\examples{
\dontrun{
config <- load_brand("block")
}
}
