% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{print.goose_cache_stats}
\alias{print.goose_cache_stats}
\title{Print Cache Statistics}
\usage{
\method{print}{goose_cache_stats}(x, ...)
}
\arguments{
\item{x}{Cache statistics object}

\item{...}{Additional arguments}
}
\description{
Print Cache Statistics
}
