% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_assistant.R
\name{print.goose_code_review}
\alias{print.goose_code_review}
\title{Print method for goose_code_review}
\usage{
\method{print}{goose_code_review}(x, ...)
}
\description{
Print method for goose_code_review
}
