% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_assistant.R
\name{print.goose_palette}
\alias{print.goose_palette}
\title{Print method for goose_palette}
\usage{
\method{print}{goose_palette}(x, ...)
}
\arguments{
\item{x}{goose_palette object}

\item{...}{additional arguments (unused)}
}
\description{
Print method for goose_palette
}
