% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_integration.R
\name{print.goose_session}
\alias{print.goose_session}
\title{Print method for goose_session}
\usage{
\method{print}{goose_session}(x, ...)
}
\arguments{
\item{x}{goose_session object}

\item{...}{additional arguments (unused)}
}
\description{
Print method for goose_session
}
