% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{print.goose_template}
\alias{print.goose_template}
\title{Print Template}
\usage{
\method{print}{goose_template}(x, ...)
}
\arguments{
\item{x}{Template object}

\item{...}{Additional arguments}
}
\description{
Print Template
}
