% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{remove_entry_from_lines}
\alias{remove_entry_from_lines}
\title{Helper: Remove Entry from Lines}
\usage{
remove_entry_from_lines(lines, name)
}
\description{
Helper: Remove Entry from Lines
}
\keyword{internal}
