% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{remove_memory_entry}
\alias{remove_memory_entry}
\title{Helper: Remove Memory Entry from Metadata}
\usage{
remove_memory_entry(metadata_path, name)
}
\description{
Helper: Remove Memory Entry from Metadata
}
\keyword{internal}
