% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goose_survey_tools.R
\name{smart_extract}
\alias{smart_extract}
\title{Smart Extract Variable Name from Question}
\usage{
smart_extract(question, max_length)
}
\description{
Smart Extract Variable Name from Question
}
\keyword{internal}
