% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{validate_brand}
\alias{validate_brand}
\title{Validate brand configuration}
\usage{
validate_brand(config)
}
\arguments{
\item{config}{Brand configuration list}
}
\value{
Logical indicating if configuration is valid
}
\description{
Validate brand configuration
}
