% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{goose_async_timeout}
\alias{goose_async_timeout}
\title{Async Query with Timeout}
\usage{
goose_async_timeout(query, timeout = 30)
}
\arguments{
\item{query}{The query to execute}

\item{timeout}{Timeout in seconds}
}
\value{
Response or timeout error
}
\description{
Execute query with timeout protection
}
