% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gooseR-package.R
\docType{package}
\name{gooseR-package}
\alias{gooseR}
\alias{gooseR-package}
\title{gooseR: R Integration for 'Goose' AI}
\description{
Seamless integration between R and 'Goose' AI capabilities including memory management, visualization enhancements, and workflow automation. Save R objects to 'Goose' memory, apply Block branding to visualizations, and manage data science project workflows. For more information about 'Goose' AI, see \url{https://github.com/block/goose}.
}
\author{
\strong{Maintainer}: Brandon Theriault \email{btheriault@block.xyz}

}
\keyword{internal}
