% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_assistant.R
\name{print.goose_error_explanation}
\alias{print.goose_error_explanation}
\title{Print method for goose_error_explanation}
\usage{
\method{print}{goose_error_explanation}(x, ...)
}
\arguments{
\item{x}{goose_error_explanation object}

\item{...}{additional arguments (unused)}
}
\description{
Print method for goose_error_explanation
}
